fastapi/requirements.txt
fastapi
uvicorn
redis==5.*
numpy
python-dotenv
pydantic
ollama
openai>=1.40.0
